compile warnings

+++ src/burn/devices/slapstic.cpp
@@ -867,7 +867,7 @@ static INT32 alt2_kludge(UINT32 /*offset*/)
 		static const SekRegister SekRegs1[8] = { SEK_REG_A0, SEK_REG_A1, SEK_REG_A2, SEK_REG_A3, SEK_REG_A4, SEK_REG_A5, SEK_REG_A6, SEK_REG_A7 };
 
 		/* first verify that the prefetched PC matches the first alternate */
-		if (MATCHES_MASK_VALUE(SekGetPC(-1) >> 1, slapstic.alt1))
+		if (MATCHES_MASK_VALUE((INT32)(SekGetPC(-1) >> 1), slapstic.alt1))
 		{
 			/* now look for a move.w (An),(An) or cmpm.w (An)+,(An)+ */
 			UINT16 opcode = SekFetchWord((SekGetPC(-1) - 4) & 0xffffff); // IQ_132 check this!
